backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Django==4.0
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dparse==0.5.1
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pbr==5.8.1
platformdirs==2.4.1
pycodestyle==2.8.0
pydot==1.4.2
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
urllib3==1.26.8
whitenoise==6.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.1